// src/lib/temple/types.ts
export type TempleAccount =
  | TempleHDAccount
  | TempleImportedAccount
  | TempleLedgerAccount
  | TempleManagedKTAccount
  | TempleWatchOnlyAccount;

export enum DerivationType {
  ED25519 = 0,
  SECP256K1 = 1,
  P256 = 2
}

export interface TempleLedgerAccount extends TempleAccountBase {
  type: TempleAccountType.Ledger;
  derivationPath: string;
}

export interface TempleImportedAccount extends TempleAccountBase {
  type: TempleAccountType.Imported;
}

export interface TempleHDAccount extends TempleAccountBase {
  type: TempleAccountType.HD;
  hdIndex: number;
}

export interface TempleManagedKTAccount extends TempleAccountBase {
  type: TempleAccountType.ManagedKT;
  chainId: string;
  owner: string;
}

export interface TempleWatchOnlyAccount extends TempleAccountBase {
  type: TempleAccountType.WatchOnly;
  chainId?: string;
}

export interface TempleAccountBase {
  type: TempleAccountType;
  name: string;
  publicKeyHash: string;
  isActivated?: boolean;
  hdIndex?: number;
  derivationPath?: string;
  derivationType?: DerivationType;
}

export enum TempleAccountType {
  HD,
  Imported,
  Ledger,
  ManagedKT,
  WatchOnly
}

export interface TempleNetwork {
  id: string;
  name: string;
  nameI18nKey?: string;
  description: string;
  descriptionI18nKey?: string;
  type: TempleNetworkType;
  rpcBaseURL: string;
  color: string;
  disabled: boolean;
  hidden?: boolean;
}

export type TempleNetworkType = 'main' | 'test';

export interface TempleSettings {
  customNetworks?: TempleNetwork[];
  lambdaContracts?: Record<string, string>;
  contacts?: TempleContact[];
}

export enum TempleSharedStorageKey {
  DAppEnabled = 'dappenabled',
  LocaleCode = 'localecode',
  UseLedgerLive = 'useledgerlive',
  PasswordAttempts = 'passwordAttempts',
  TimeLock = 'timelock'
}

export type TempleDAppSessions = Record<string, TempleDAppSession>;

export interface TempleContact {
  address: string;
  name: string;
  addedAt?: number;
  accountInWallet?: boolean;
}